requirements.txt
fastapi
uvicorn[standard]
httpx
langgraph
langchain-core
langchain-openai
python-dotenv
sentence-transformers==2.7.0